requirements.txt
flask==2.3.3
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
supabase==2.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
